requirements.txt
requests
pillow
pillow-avif-plugin
json5
